src/main/java/leetcode/problems/remove_linked_list_elements/Solution.java
package leetcode.problems.remove_linked_list_elements;

import leetcode.struct.ListNode;

class Solution {
    public ListNode removeElements(ListNode head, int val) {
        ListNode pre = new ListNode(0);
        pre.next = head;
        
        ListNode cur = pre;
        while (cur.next != null) {
            if (cur.next.val == val) {
                cur.next = cur.next.next;
            } else {
                cur = cur.next;
            }
        }
        return pre.next;
    }
    
    
}